+++ lib/utils/app_constants.dart
@@ -1,4 +1,9 @@
 class AppConstants {
   static const String author = "Karnaukhov Oleh";
   static const String supportEmail = "karnauhov.oleg@gmail.com";
+  static const languages = {
+    'en': 'English',
+    'uk': 'Українська',
+    'ru': 'Русский',
+  };
 }