+++ lib/utils/stub.dart
@@ -1,0 +1,3 @@
+String getPlatformLanguage() {
+  return 'en';
+}